+++ packages/segmented-control/src/SegmentedControl/SegmentedControl.tsx
@@ -29,8 +29,12 @@ import {
   optionsWrapperStyle,
   selectionIndicatorStyle,
   wrapperStyle,
-} from './styles';
-import { SegmentedControlProps, Size } from './types';
+} from './SegmentedControl.styles';
+import {
+  DeprecatedSize,
+  SegmentedControlProps,
+  Size,
+} from './SegmentedControl.types';
 
 /**
  * Segmented controls act as a toggle between a current state and related states, often changing the view of information within a single page.
@@ -42,7 +46,7 @@ export const SegmentedControl = forwardRef<
   {
     children,
     name: nameProp,
-    size = Size.Default,
+    size: sizeProp = Size.Default,
     darkMode: darkModeProp,
     defaultValue,
     value: controlledValue,
@@ -302,6 +306,9 @@ export const SegmentedControl = forwardRef<
     },
     [renderedChildren],
   );
+
+  // @ts-expect-error deprecated size "small" no longer supported, but we want to gracefully degrade the prop value
+  const size = sizeProp === DeprecatedSize.Small ? Size.XSmall : sizeProp;
 
   /**
    * Return
