+++ packages/segmented-control/src/SegmentedControl/SegmentedControl.types.ts
@@ -1,7 +1,13 @@
 import { HTMLElementProps } from '@leafygreen-ui/lib';
 
-export const Size = {
+export const DeprecatedSize = {
   Small: 'small',
+} as const;
+
+export type DeprecatedSize = typeof DeprecatedSize[keyof typeof DeprecatedSize];
+
+export const Size = {
+  XSmall: 'xsmall',
   Default: 'default',
   Large: 'large',
 } as const;
@@ -18,7 +24,7 @@ export interface SegmentedControlProps
   children: React.ReactNode;
 
   /**
-   * Defines the size of the segmented control. Can be either `small`, `default`, or `large`
+   * Defines the size of the segmented control. Can be either `xsmall`, `default`, or `large`
    */
   size?: Size;
