+++ packages/segmented-control/src/SegmentedControlOption/SegmentedControlOption.types.ts
