+++ packages/segmented-control/src/index.ts
@@ -1,4 +1,4 @@
 export { SegmentedControl } from './SegmentedControl/SegmentedControl';
-export type { SegmentedControlProps } from './SegmentedControl/types';
+export type { SegmentedControlProps } from './SegmentedControl/SegmentedControl.types';
 export { SegmentedControlOption } from './SegmentedControlOption/SegmentedControlOption';
-export type { SegmentedControlOptionProps } from './SegmentedControlOption/types';
+export type { SegmentedControlOptionProps } from './SegmentedControlOption/SegmentedControlOption.types';
